+++ cmd/export.go
@@ -79,7 +79,7 @@ func export(cliCtx *cli.Context) error {
 		return cli.NewExitError("Could not create marketplace client: "+err.Error(), -1)
 	}
 
-	r, err := clients.FetchResources(ctx, marketplace, teamID)
+	r, err := clients.FetchResources(ctx, marketplace, teamID, true)
 	if err != nil {
 		return cli.NewExitError("Could not retrieve resources: "+err.Error(), -1)
 	}