+++ cmd/team.go
@@ -205,7 +205,7 @@ func leaveTeamCmd(cliCtx *cli.Context) error {
 		return err
 	}
 
-	memberships, err := clients.FetchMemberships(ctx, identityClient)
+	memberships, err := clients.FetchMemberships(ctx, identityClient, false)
 	if err != nil {
 		return cli.NewExitError(fmt.Sprintf("Failed to fetch user memberships: %s", err), -1)
 	}
@@ -357,7 +357,7 @@ func loadIdentityClient() (*client.Identity, error) {
 // fetchTeams retrieves all user's team and prompt to select which team the cmd
 // will be applied to.
 func selectTeam(ctx context.Context, teamName string, identityClient *client.Identity) (*iModels.Team, error) {
-	teams, err := clients.FetchTeams(ctx, identityClient)
+	teams, err := clients.FetchTeams(ctx, identityClient, false)
 	if err != nil {
 		return nil, cli.NewExitError(fmt.Sprintf("Failed to fetch list of teams: %s", err), -1)
 	}