+++ clients/resources.go
@@ -4,6 +4,9 @@ import (
 	"context"
 
 	"github.com/manifoldco/go-manifold"
+
+	"github.com/manifoldco/manifold-cli/prompts"
+
 	mClient "github.com/manifoldco/manifold-cli/generated/marketplace/client"
 	"github.com/manifoldco/manifold-cli/generated/marketplace/client/resource"
 	mModels "github.com/manifoldco/manifold-cli/generated/marketplace/models"
@@ -13,7 +16,12 @@ import (
 )
 
 // FetchOperations returns the resources for the authenticated user
-func FetchOperations(ctx context.Context, c *pClient.Provisioning, teamID *manifold.ID) ([]*pModels.Operation, error) {
+func FetchOperations(ctx context.Context, c *pClient.Provisioning, teamID *manifold.ID, shouldSpin bool) ([]*pModels.Operation, error) {
+	if shouldSpin {
+		spin := prompts.NewSpinner("Fetching operations")
+		spin.Start()
+		defer spin.Stop()
+	}
 	res, err := c.Operation.GetOperations(
 		operation.NewGetOperationsParamsWithContext(ctx), nil)
 	if err != nil {
@@ -31,7 +39,12 @@ func FetchOperations(ctx context.Context, c *pClient.Provisioning, teamID *manif
 }
 
 // FetchResources returns the resources for the authenticated user
-func FetchResources(ctx context.Context, c *mClient.Marketplace, teamID *manifold.ID) ([]*mModels.Resource, error) {
+func FetchResources(ctx context.Context, c *mClient.Marketplace, teamID *manifold.ID, shouldSpin bool) ([]*mModels.Resource, error) {
+	if shouldSpin {
+		spin := prompts.NewSpinner("Fetching resources")
+		spin.Start()
+		defer spin.Stop()
+	}
 	res, err := c.Resource.GetResources(
 		resource.NewGetResourcesParamsWithContext(ctx), nil)
 	if err != nil {