+++ clients/teams.go
@@ -6,6 +6,7 @@ import (
 	iClient "github.com/manifoldco/manifold-cli/generated/identity/client"
 	"github.com/manifoldco/manifold-cli/generated/identity/client/team"
 	iModels "github.com/manifoldco/manifold-cli/generated/identity/models"
+	"github.com/manifoldco/manifold-cli/prompts"
 )
 
 // TeamMembersCount groups a team name with the amount of members the team has
@@ -15,9 +16,14 @@ type TeamMembersCount struct {
 }
 
 // FetchTeams returns the teams for the authenticated user
-func FetchTeams(ctx context.Context, c *iClient.Identity) ([]*iModels.Team, error) {
-	params := team.NewGetTeamsParamsWithContext(ctx)
-	res, err := c.Team.GetTeams(params, nil)
+func FetchTeams(ctx context.Context, c *iClient.Identity, shouldSpin bool) ([]*iModels.Team, error) {
+	if shouldSpin {
+		spin := prompts.NewSpinner("Fetching teams")
+		spin.Start()
+		defer spin.Stop()
+	}
+	res, err := c.Team.GetTeams(
+		team.NewGetTeamsParamsWithContext(ctx), nil)
 	if err != nil {
 		return nil, err
 	}
@@ -25,7 +31,12 @@ func FetchTeams(ctx context.Context, c *iClient.Identity) ([]*iModels.Team, erro
 }
 
 // FetchTeamMembers returns a list of members profile from a team.
-func FetchTeamMembers(ctx context.Context, id string, c *iClient.Identity) ([]*iModels.MemberProfile, error) {
+func FetchTeamMembers(ctx context.Context, id string, c *iClient.Identity, shouldSpin bool) ([]*iModels.MemberProfile, error) {
+	if shouldSpin {
+		spin := prompts.NewSpinner("Fetching memberships")
+		spin.Start()
+		defer spin.Stop()
+	}
 	params := team.NewGetTeamsIDMembersParamsWithContext(ctx)
 	params.SetID(id)
 	res, err := c.Team.GetTeamsIDMembers(params, nil)
@@ -37,8 +48,13 @@ func FetchTeamMembers(ctx context.Context, id string, c *iClient.Identity) ([]*i
 
 // FetchTeamsMembersCount returns a list of all user teams with their names and
 // number of members.
-func FetchTeamsMembersCount(ctx context.Context, c *iClient.Identity) ([]TeamMembersCount, error) {
-	teams, err := FetchTeams(ctx, c)
+func FetchTeamsMembersCount(ctx context.Context, c *iClient.Identity, shouldSpin bool) ([]TeamMembersCount, error) {
+	if shouldSpin {
+		spin := prompts.NewSpinner("Fetching memberships")
+		spin.Start()
+		defer spin.Stop()
+	}
+	teams, err := FetchTeams(ctx, c, false)
 	if err != nil {
 		return nil, err
 	}
@@ -53,7 +69,7 @@ func FetchTeamsMembersCount(ctx context.Context, c *iClient.Identity) ([]TeamMem
 		name := string(t.Body.Name)
 
 		go func() {
-			members, err := FetchTeamMembers(ctx, id, c)
+			members, err := FetchTeamMembers(ctx, id, c, false)
 
 			if err != nil {
 				fail <- err
@@ -81,7 +97,12 @@ func FetchTeamsMembersCount(ctx context.Context, c *iClient.Identity) ([]TeamMem
 }
 
 // FetchMemberships returns all memberships for the authenticated user
-func FetchMemberships(ctx context.Context, c *iClient.Identity) ([]iModels.TeamMembership, error) {
+func FetchMemberships(ctx context.Context, c *iClient.Identity, shouldSpin bool) ([]iModels.TeamMembership, error) {
+	if shouldSpin {
+		spin := prompts.NewSpinner("Fetching memberships")
+		spin.Start()
+		defer spin.Stop()
+	}
 	params := team.NewGetMembershipsParamsWithContext(ctx)
 	res, err := c.Team.GetMemberships(params, nil)