+++ cmd/config.go
@@ -11,6 +11,7 @@ import (
 	"github.com/manifoldco/manifold-cli/clients"
 	"github.com/manifoldco/manifold-cli/config"
 	"github.com/manifoldco/manifold-cli/middleware"
+	"github.com/manifoldco/manifold-cli/prompts"
 
 	"github.com/manifoldco/manifold-cli/generated/marketplace/client/credential"
 	"github.com/manifoldco/manifold-cli/generated/marketplace/models"
@@ -78,7 +79,7 @@ func patchConfig(cliCtx *cli.Context, req map[string]*string) error {
 		return cli.NewExitError("Could not create marketplace client: "+err.Error(), -1)
 	}
 
-	resources, err := clients.FetchResources(ctx, marketplace, teamID)
+	resources, err := clients.FetchResources(ctx, marketplace, teamID, false)
 	if err != nil {
 		return cli.NewExitError("Could not retrieve resources: "+err.Error(), -1)
 	}
@@ -99,6 +100,9 @@ func patchConfig(cliCtx *cli.Context, req map[string]*string) error {
 		return cli.NewExitError("Config can only be set on custom resources", -1)
 	}
 
+	spin := prompts.NewSpinner("Setting resource config")
+	spin.Start()
+	defer spin.Stop()
 	_, err = marketplace.Credential.PatchResourcesIDConfig(&credential.PatchResourcesIDConfigParams{
 		ID:      resource.ID.String(),
 		Body:    req,