+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Add `billing redeem` command to redeem coupon codes
+- Add `teams list` command to list all teams and # of members
 - Conflict response when redeeming coupons
 
 ### Fixed
@@ -31,7 +33,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add 'switch' to switch Teams, so interact with a team's resources
 - Add --team and --me flags for operating under your account, or a team's
 - Add --team and --me flags for billing
-- Add `billing redeem` command to redeem coupon codes
 
 ### Fixed